# docs/swagger.yaml
definitions:
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterRequest:
    properties:
      balance:
        type: number
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 20
        minLength: 4
        type: string
      surname:
        type: string
    required:
    - balance
    - email
    - name
    - password
    - surname
    type: object
  request.TransferRequest:
    properties:
      amount:
        type: number
      receiver:
        type: string
      sender:
        type: string
    required:
    - amount
    - receiver
    - sender
    type: object
  response.LoginResponse:
    properties:
      token:
        type: string
    type: object
  response.RegisterResponse:
    properties:
      account_number:
        type: string
      balance:
        type: number
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      surname:
        type: string
    type: object
  response.TransferResponse:
    properties:
      fee:
        type: number
      message:
        type: string
      receiver_balance:
        type: number
      sender_balance:
        type: number
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login a user with the provided credentials
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login a user
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details
      parameters:
      - description: Register Request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /transfer:
    post:
      consumes:
      - application/json
      description: Transfer money from one account to another
      parameters:
      - description: Transfer Request
        in: body
        name: transferRequest
        required: true
        schema:
          $ref: '#/definitions/request.TransferRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TransferResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Transfer money between accounts
      tags:
      - transaction
swagger: "2.0"